requirements.txt
pytest==8.3.5
pytest-xdist==3.6.1
pytest-asyncio==0.25.2
allure-pytest==2.13.5

requests==2.32.3
jsonschema==4.23.0

psycopg2-binary==2.9.10

kafka-python==2.0.2

locust==2.29.1

selenium==4.20.0

playwright==1.49.1

kubernetes==29.0.0

elasticsearch==8.17.1

tenacity==9.0.0
pydantic==2.10.6
python-dotenv==1.0.1